apps/balance-gateway/pyproject.toml
[tool.poetry]
name = "balance-gateway"
version = "0.1.0"
description = "FastAPI REST gateway that translates HTTP requests to gRPC calls"
authors = []
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
uvicorn = { extras = ["standard"], version = "^0.29.0" }
grpcio = "^1.62.0"
protobuf = "^4.25.0"

[tool.poetry.group.dev.dependencies]
grpcio-tools = "^1.62.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
